// src/types/key.ts
export enum Key{
    CE='CE',
    C='C',
    BACKSPACE = '<-',
    DIVIDE = '/',
    MULTIPLY = '*',
    PLUS = '+',
    MINUS = '-',
    ZERO = '0',
    ONE = '1',
    TWO = '2',
    THREE = '3',
    FOUR = '4',
    FIVE = '5',
    SIX = '6',
    SEVEN = '7',
    EIGHT = '8',
    NINE = '9',
    PLUS_MINUS = '+/-',
    DOT = '.',
    NEGATIVE = '-',
    EQUAL = '='
}